Struct secret_cosmwasm_std::Binary

pub struct Binary(pub Vec<u8>);

Binary is a wrapper around Vec to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.

This is only needed as serde-json-{core,wasm} has a horrible encoding for Vec

Tuple Fields
0: Vec<u8>
Implementations

impl Binary

pub fn from_base64(encoded: &str) -> StdResult<Self>

take an (untrusted) string and decode it into bytes. fails if it is not valid base64

pub fn to_base64(&self) -> String

encode to base64 string (guaranteed to be success as we control the data inside). this returns normalized form (with trailing = if needed)

pub fn as_slice(&self) -> &[u8]

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool